pyproject.toml
[tool.poetry]
name = "ssl-expiry-checker-azure-func"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pyopenssl = "^25.0.0"
requests = "^2.32.3"
jinja2 = "^3.1.5"
azure-functions = "^1.21.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
